equihash: Move helper methods for minimal reprs into a submodule
