+++ components/equihash/src/lib.rs
@@ -20,6 +20,7 @@
 // Catch documentation errors caused by code changes.
 #![deny(rustdoc::broken_intra_doc_links)]
 
+mod minimal;
 mod params;
 mod verify;
